+++ js/main.js
@@ -1,3 +1,2 @@
-import {createAnnouncements} from './util.js';
-const ANNOUNCEMENTS_NUMBER = 10;
-createAnnouncements(ANNOUNCEMENTS_NUMBER);
+import './data.js';
+import './popup.js';
